+++ crates/zng-wgt-scroll/src/node.rs
@@ -601,9 +601,22 @@ pub fn zoom_commands_node(child: impl IntoUiNode) -> UiNode {
                     WIDGET.layout();
                 });
             } else if let Some(args) = ZOOM_TO_FIT_CMD.scoped(scope).on(update) {
-                args.handle_enabled(&zoom_to_fit, |_| {
-                    let scale = fit_scale();
-                    SCROLL.chase_zoom(|_| scale);
+                args.handle_enabled(&zoom_to_fit, |args| {
+                    let mut scale = fit_scale();
+                    if let Some(p) = ZoomToFitRequest::from_args(args) {
+                        if p.scale_down {
+                            scale = scale.min(1.fct());
+                        }
+                        if p.skip_animation {
+                            SMOOTH_SCROLLING_VAR.with_context_var(zng_var::ContextInitHandle::current(), false, || {
+                                SCROLL.chase_zoom(|_| scale);
+                            });
+                        } else {
+                            SCROLL.chase_zoom(|_| scale);
+                        }
+                    } else {
+                        SCROLL.chase_zoom(|_| scale);
+                    }
                 });
             } else if let Some(args) = ZOOM_RESET_CMD.scoped(scope).on(update) {
                 args.handle_enabled(&zoom_reset, |_| {