+++ crates/zng/src/scroll.rs
@@ -85,6 +85,6 @@ pub mod cmd {
         PAGE_DOWN_CMD, PAGE_LEFT_CMD, PAGE_RIGHT_CMD, PAGE_UP_CMD, SCROLL_DOWN_CMD, SCROLL_LEFT_CMD, SCROLL_RIGHT_CMD,
         SCROLL_TO_BOTTOM_CMD, SCROLL_TO_CMD, SCROLL_TO_LEFTMOST_CMD, SCROLL_TO_RIGHTMOST_CMD, SCROLL_TO_TOP_CMD, SCROLL_UP_CMD,
         ScrollRequest, ScrollToMode, ScrollToRequest, ScrollToTarget, ZOOM_IN_CMD, ZOOM_OUT_CMD, ZOOM_RESET_CMD, ZOOM_TO_FIT_CMD,
-        scroll_to, scroll_to_zoom,
+        ZoomToFitRequest, scroll_to, scroll_to_zoom,
     };
 }